# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ExecuTorch Bencher API",
    "description": "Internal Android LLM benchmark collection and analysis service. Provides HTTP endpoints to ingest, store, and retrieve benchmark run results, content-addressed artifacts, and registered model assets.",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/artifacts/{id}/content": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Stream an artifact's content, decompressing transparently if it is\nstored Zstandard-compressed. The response body is the raw content with\nits stored media type - never a JSON-encoded string.",
        "operationId": "getArtifactContent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The artifact's identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The artifact's content, streamed with its stored media type (or `application/octet-stream` if none was recorded).",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/BinaryBody"
                }
              }
            }
          },
          "404": {
            "description": "No artifact exists with this ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "The database record exists but the file is missing on disk.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{id}/download": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Like `content`, but sets `Content-Disposition: attachment` with a safe\nfilename derived from the artifact's kind and content hash.",
        "operationId": "downloadArtifact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The artifact's identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The artifact's content as an attachment download.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/BinaryBody"
                }
              }
            }
          },
          "404": {
            "description": "No artifact exists with this ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "The database record exists but the file is missing on disk.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/runs/{id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Fetch a run by ID, including its outcome and the artifacts/model it\nreferenced. See `specs/ingestion-service` - \"Run responses expose\nviewable artifact metadata\".",
        "operationId": "getRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The run's identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The run was found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "404": {
            "description": "No run exists with this ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "description": "The `error` object inside an [`ApiErrorResponse`].",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable, machine-readable failure category. Current values: `invalid_request`,\n`not_found`, `artifact_file_missing`, `payload_too_large`, `conflict`,\n`not_implemented`, `internal_error`. This field is an open string, not a\nclosed enum - new categories may be added as the API grows.",
            "example": "invalid_request"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the failure. Not guaranteed stable across\nversions or releases - match on `code`, not this text.",
            "example": "gpu_clock_mhz must be greater than zero"
          },
          "details": {
            "description": "Optional structured context about the failure, e.g. which field was invalid."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque identifier correlating this error with server-side logs, when available."
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "The consistent JSON envelope every rejected or failed request responds with.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        }
      },
...
      "BinaryBody": {
        "type": "string",
        "format": "binary",
        "description": "OpenAPI-only marker for a raw, unstructured binary body (streamed as\n`application/octet-stream`, not JSON and not a multipart form)."
      },
...
    }
  },
...
}